Refactor Stage 3
